refactor: extract sparse-tree helpers to rln-layouts
subtree_node_offset and read_sparse_node were duplicated across the guest
merkle_tree module, the host merkle_tree::client module, and lez-rln-ffi.
All three implementations were byte-identical sparse-storage primitives,
kept in sync by convention.

Moves them into a new rln_layouts::sparse module (no_std compatible) and
has each of the three callers import from there. No behavioral change.
