+++ lez-rln/methods/guest/src/merkle_tree.rs
@@ -33,66 +33,8 @@ pub use rln_layouts::{
     TREE_DEPTH, TOP_DEPTH, BOTTOM_DEPTH, SUBTREE_LEAVES,
     OFFSET_DEPTH, OFFSET_NEXT_INDEX, OFFSET_ROOT, OFFSET_ROOT_HISTORY, ROOT_HISTORY_SIZE,
     OFFSET_CACHED_NODES, OFFSET_TOP_TREE_DATA,
+    read_sparse_node, subtree_node_offset,
 };
-
-// ============================================================================
-// Sparse Node Storage
-// ============================================================================
-
-/// Compute the BFS offset for a node at (level, index_within_level).
-/// Used as the key in sparse node storage.
-#[inline]
-fn subtree_node_offset(level: usize, index: usize) -> usize {
-    ((1 << level) - 1) + index
-}
-
-/// Read a 32-byte hash from sparse tree node storage.
-///
-/// Format: `[count(u16le), entries...]` where each entry is `[offset(u16le), hash(32)]`.
-/// Entries are sorted by offset for binary search. Returns cached_default if not found.
-fn read_sparse_node(
-    data: &[u8],
-    level: usize,
-    index: usize,
-    cached_default: &[u8; 32],
-) -> [u8; 32] {
-    if data.len() < 2 {
-        return *cached_default;
-    }
-    let count = u16::from_le_bytes(data[0..2].try_into().unwrap()) as usize;
-    if count == 0 {
-        return *cached_default;
-    }
-    let target = subtree_node_offset(level, index) as u16;
-
-    // Binary search
-    let mut lo = 0usize;
-    let mut hi = count;
-    while lo < hi {
-        let mid = lo + (hi - lo) / 2;
-        let entry_start = 2 + mid * 34;
-        let entry_offset = u16::from_le_bytes(
-            data[entry_start..entry_start + 2].try_into().unwrap(),
-        );
-        if entry_offset < target {
-            lo = mid + 1;
-        } else {
-            hi = mid;
-        }
-    }
-
-    if lo < count {
-        let entry_start = 2 + lo * 34;
-        let entry_offset = u16::from_le_bytes(
-            data[entry_start..entry_start + 2].try_into().unwrap(),
-        );
-        if entry_offset == target {
-            return data[entry_start + 2..entry_start + 34].try_into().unwrap();
-        }
-    }
-
-    *cached_default
-}
 
 /// Write a 32-byte hash into sparse tree node storage.
 ///
